Move Lrama::Code to Lrama::Grammar::Code
