+++ lib/lrama.rb
@@ -1,5 +1,4 @@
 require "lrama/bitmap"
-require "lrama/code"
 require "lrama/command"
 require "lrama/context"
 require "lrama/digraph"
